Add login function
Voir ajout enseignant, gender too long

+++ src/index.php
@@ -7,6 +7,38 @@
 
     // Récupérer la liste des enseignants depuis la base de données
     $enseignants =  $db->getAllTeachers();
+
+    $users = [];
+    $isUserConnected = 0;
+
+    // Récupération du login et password saisi par l'utilisateur
+    // A noter que ?? retourne "" si $_POST["useLogin"] est null 
+    $useLogin = $_POST["useLogin"] ?? "";
+    $userPassword = $_POST["usePassword"] ?? "";
+
+    // Création d'une instance PDO pour se connecter à la base de données
+    // ATTENTION si vous n'utilisez pas Docker mais uWamp
+    // vous devez supprimer le port du host 
+    // "mysql:host=localhost;dbname=challengeInjectionSQL;charset=utf8"
+    $connector = new PDO("mysql:host=localhost:6033;dbname=db_nickname;charset=utf8", 'root', 'root');
+
+    // Création de la requête SQL permettant de récupérer les informations de l'utilisateur
+    // à partir du login et password saisis par l'utilisateur
+    $query = "SELECT * FROM t_user where useLogin = :useLogin and usePassword = :usePassword";
+
+    // Exécution de la requête. A noter l'utilisation de la méthode ->query()
+    $req = $connector->prepare($query);
+    $req -> bindValue('useLogin', $useLogin, PDO::PARAM_STR);
+    $req -> bindValue('usePassword', $userPassword, PDO::PARAM_STR);
+    $req -> execute();
+
+    // On convertit le résultat de la requête en tableau
+    $user = $req->fetchALL(PDO::FETCH_ASSOC);
+    // Si le tableau 'user' n'est pas vide, cela signifie que l'utilisateur a bien été trouvé en DB
+    if ($user) {
+        $isUserConnected = 1;
+    }
+
 ?>
 
 <!DOCTYPE html>
@@ -28,13 +60,25 @@
                 <h1>Surnom des enseignants</h1>
             </div>
             <div class="login-container">
-                <form action="#" method="post">
+                <form action="index.php" method="post">
                     <label for="user"> </label>
-                    <input type="text" name="user" id="user" placeholder="Login">
+                    <input type="text" name="useLogin" id="user" placeholder="Login">
                     <label for="password"> </label>
-                    <input type="password" name="password" id="password" placeholder="Mot de passe">
+                    <input type="password" name="usePassword" id="password" placeholder="Mot de passe">
                     <button type="submit" class="btn btn-login">Se connecter</button>
                 </form>
+            </div>
+            <div class="block">
+            <h2>Utilisateur connecté ? :
+                <?php if ($isUserConnected === 1) {
+                    echo "Oui";
+                } else {
+                    echo "Non";
+                }  ?>
+            </h2>
+            <?php if ($isUserConnected) {
+                echo "<h2>Bonjour " . $user[0]["useLogin"] . " </h2>";
+            } ?>
             </div>
         </div>
         <nav>
@@ -60,16 +104,24 @@
                     <tr>
                         <td><?php echo $enseignant['teaFirstname']; ?></td>
                         <td><?php echo $enseignant['teaNickname']; ?></td>
+                        <!-- Affichage des boutons d'édition et de suppression -->
                         <td class="containerOptions">
-                            <a href="./updateTeacher.php?idTeacher=<?= $enseignant["idTeacher"]; ?>">
-                                <img src="./img/edit.png" alt="edit">
-                            </a>
-                            <a href="javascript:confirmDelete(<?= $enseignant["idTeacher"]; ?>)">
-                                <img src="./img/delete.png" alt="delete">
-                            </a>
-                            <a href="./detailTeacher.php?idTeacher=<?= $enseignant["idTeacher"]; ?>">
+                            <?php if ($isUserConnected === 1) : ?>
+                                <a href="./updateTeacher.php?idTeacher=<?= $enseignant["idTeacher"]; ?>">
+                                    <img src="./img/edit.png" alt="edit">
+                                </a>
+                                <a href="javascript:confirmDelete(<?= $enseignant["idTeacher"]; ?>)">
+                                    <img src="./img/delete.png" alt="delete">
+                                </a>
+                                <a href="./detailTeacher.php?idTeacher=<?= $enseignant["idTeacher"]; ?>">
                                 <img src="./img/detail.png" alt="detail">
-                            </a>
+                                </a>
+                            <?php endif; ?>
+                            <?php if ($isUserConnected !== 1) : ?>
+                                <a href="./detailTeacher.php?idTeacher=<?= $enseignant["idTeacher"]; ?>">
+                                <img src="./img/detail.png" alt="detail">
+                                </a>
+                            <?php endif; ?>
                         </td>
                     </tr>
                 <?php endforeach; ?>
